src/ingest/nfl/team_weeks.py
"""
NFL Team Week data collection module.

This module handles fetching and processing NFL team performance data on a weekly basis.
It includes functionality for:
- Fetching team statistics
- Processing team performance data
- Caching results
"""

import os
from typing import Dict, List
from datetime import datetime
import requests
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

class NFLTeamWeeksIngest:
    """Handles NFL team weekly data ingestion."""
    
    def __init__(self):
        """Initialize the NFL team weeks ingest with API configuration."""
        self.base_url = os.getenv("NFL_API_BASE_URL")
        self.api_key = os.getenv("NFL_API_KEY")
        self.cache_dir = os.getenv("NFL_CACHE_DIR", "data/raw/team_weeks")
        
    def fetch_team_week(self, season: int, week: int, team_id: str) -> Dict:
        """
        Fetch team performance data for a specific week.
        
        Args:
            season: The NFL season year
            week: The week number
            team_id: The NFL team identifier
            
        Returns:
            Dict containing the team's weekly performance data
        """
        raise NotImplementedError("Method needs to be implemented")
    
    def process_team_data(self, raw_data: Dict) -> Dict:
        """
        Process raw team data into a standardized format.
        
        Args:
            raw_data: Raw team data from the API
            
        Returns:
            Processed team performance data
        """
        raise NotImplementedError("Method needs to be implemented") 