requirements.txt
pandas>=2.0.0
selenium>=4.0.0
beautifulsoup4>=4.9.0
python-dotenv>=0.19.0
webdriver-manager>=3.8.0 